[EI-136] [api] rename: types -> type-definition.d.ts
