+++ api/src/numerical-guidance/api/dto/get-fluctuatingIndicator.dto.ts
@@ -2,7 +2,7 @@ import { IsInt, IsString } from 'class-validator';
 import { Transform } from 'class-transformer';
 import { IsInterval } from '../../../utils/validation/is.interval.validation';
 import { ApiProperty } from '@nestjs/swagger';
-import { Market } from 'src/utils/type/types';
+import { Market } from 'src/utils/type/type-definition';
 import { IsMarket } from 'src/utils/validation/is.market.validation';
 
 type Interval = 'day' | 'week' | 'month' | 'year';
