+++ api/src/numerical-guidance/application/query/get-indicator-list/indicator-list.dto.ts
@@ -1,4 +1,4 @@
-import { IndicatorType, Market } from 'src/utils/type/types';
+import { IndicatorType, Market } from 'src/utils/type/type-definition';
 
 export class IndicatorListDto {
   indicatorList: Indicator[];
